refactor: renamed prediction buy buttion
